Move again SecureShareContent impl
