+++ massa-models/src/endorsement.rs
@@ -1,7 +1,8 @@
 // Copyright (c) 2022 MASSA LABS <info@massa.net>
 
+use crate::denunciation::{DenunciationData, DenunciationDataSerializer};
 use crate::prehash::PreHashed;
-use crate::secure_share::{Id, SecureShare};
+use crate::secure_share::{Id, SecureShare, SecureShareContent};
 use crate::slot::{Slot, SlotDeserializer, SlotSerializer};
 use crate::{block_id::BlockId, error::ModelsError};
 use massa_hash::{Hash, HashDeserializer};
@@ -9,6 +10,7 @@ use massa_serialization::{
     DeserializeError, Deserializer, SerializeError, Serializer, U32VarIntDeserializer,
     U32VarIntSerializer, U64VarIntDeserializer, U64VarIntSerializer,
 };
+use massa_signature::PublicKey;
 use nom::error::context;
 use nom::sequence::tuple;
 use nom::Parser;
@@ -165,6 +167,25 @@ impl SecureShareEndorsement {
 
 /// Wrapped endorsement
 pub type SecureShareEndorsement = SecureShare<Endorsement, EndorsementId>;
+
+impl SecureShareContent for Endorsement {
+    fn compute_hash(
+        content: &Self,
+        content_serialized: &[u8],
+        content_creator_pub_key: &PublicKey,
+    ) -> Result<Hash, SerializeError> {
+        let de_data = DenunciationData::Endorsement((content.slot, content.index));
+        let de_data_serializer = DenunciationDataSerializer::new();
+        let mut de_data_ser = Vec::new();
+        de_data_serializer.serialize(&de_data, &mut de_data_ser)?;
+
+        let mut hash_data = Vec::new();
+        hash_data.extend(content_creator_pub_key.to_bytes());
+        hash_data.extend(de_data_ser);
+        hash_data.extend(Hash::compute_from(content_serialized).to_bytes());
+        Ok(Hash::compute_from(&hash_data))
+    }
+}
 
 /// Serializer for `Endorsement`
 #[derive(Clone)]
